requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-dotenv
django-filter
pytest
pytest-django
drf-spectacular